Split sqlite store internals
